+++ apps/scan/frontend/src/utils/config.ts
@@ -2,6 +2,7 @@ import { ConfigReader, serviceSchema } from 'common-frontend';
 import { z } from 'zod';
 
 type ScanServicesConfig = {
+  directory: z.infer<typeof serviceSchema>;
   scan: z.infer<typeof serviceSchema>;
 };
 
@@ -12,6 +13,7 @@ type ScanConfig = {
 const reader = new ConfigReader(
   z.object({
     services: z.object({
+      directory: serviceSchema,
       scan: serviceSchema,
     }),
   })