+++ cluster/images/docs/src/release_notes.rst
@@ -3,6 +3,14 @@
 Release Notes
 =============
 
+2023-10-16
+----------
+
+* Frontend updates
+
+  * The recent activity tab in Scan now looks up CNS entries for party IDs.
+    Party IDs are shown as-is if they are not found in CNS.
+
 2023-10-09
 ----------